+++ src/commons/resources/index.ts
@@ -73,6 +73,12 @@ import totalOutputUrl from "./icons/totalOutput.svg";
 import CopyIconSquareUrl from "./images/copy.svg";
 import ADALogoUrl from "./icons/ada-logo.svg";
 import fileGuardUrl from "./icons/file-guard.svg";
+import DelegationIconUrl from "./icons/delegationIcon.svg";
+import RewardsDistributionIconUrl from "./icons/rewardsDistribution.svg";
+import StakeCertificatesIconUrl from "./icons/stakeCertificateIcon.svg";
+import ProtocolUpdateIconUrl from "./icons/protocolUpdateIcon.svg";
+import InstantaneousHistoryIconUrl from "./icons/instantaneousHistoryIcon.svg";
+import MetadataIconUrl from "./icons/metadataIcon2.svg";
 
 export {
   LogoFullIcon,
@@ -149,7 +155,13 @@ export {
   txOutputIconUrl,
   totalOutputUrl,
   CopyIconSquareUrl,
-  fileGuardUrl
+  fileGuardUrl,
+  DelegationIconUrl,
+  RewardsDistributionIconUrl,
+  StakeCertificatesIconUrl,
+  ProtocolUpdateIconUrl,
+  InstantaneousHistoryIconUrl,
+  MetadataIconUrl
 };
 
 export { ReactComponent as AIcon } from "./icons/AIcon.svg";
@@ -269,3 +281,4 @@ export { ReactComponent as RewardsAvailableIcon } from "./icons/rewards-availabl
 export { ReactComponent as StatusIcon } from "./icons/status-ic.svg";
 export { ReactComponent as RedeemerPlusIcon } from "./icons/redeemer-plus.svg";
 export { ReactComponent as RedeemerArrowDownIcon } from "./icons/redeemer-arrow-down.svg";
+export { ReactComponent as IRisLogo } from "./icons/iris-logo.svg";