+++ src/components/DelegationPool/DelegationList/index.tsx
@@ -124,7 +124,7 @@ const DelegationLists: React.FC = () => {
     },
     {
       title: "Fixed Cost (A)",
-      key: "fixedCost",
+      key: "pu.fixedCost",
       minWidth: "120px",
       render: (r) => `${formatADAFull(r.feeAmount)} A`,
       sort: ({ columnKey, sortValue }) => {