+++ src/components/TabularView/StakeTab/Tabs/DelegationTab.tsx
@@ -50,7 +50,7 @@ const DelegationTab = () => {
       )
     },
     {
-      title: "Timestamp",
+      title: "Created At",
       key: "id",
       minWidth: "120px",
       render: (r) => formatDateTimeLocal(r.time),