+++ src/components/TransactionLists/index.tsx
@@ -156,7 +156,7 @@ const TransactionList: React.FC<TransactionListProps> = ({
   return (
     <Card
       data-testid="transactions-card"
-      title={pathname === "/transactions" ? "Transactions" : ""}
+      title={pathname?.includes("/transactions") ? "Transactions" : ""}
       underline={underline}
     >
       <Table
@@ -167,7 +167,7 @@ const TransactionList: React.FC<TransactionListProps> = ({
           ...pageInfo,
           total: fetchData.total,
           onChange: (page, size) => {
-            mainRef.current?.scrollTo(0, 0);
+            mainRef.current?.scrollTo({ top: 0, behavior: "smooth" });
             history.replace({ search: stringify({ page, size }) });
           },
           handleCloseDetailView: handleClose