+++ src/pages/ContractList/index.tsx
@@ -34,7 +34,7 @@ const Transactions: React.FC = () => {
   const { adaRate } = useSelector(({ system }: RootState) => system);
 
   useEffect(() => {
-    document.title = `Contracts List | Cardano Explorer`;
+    document.title = `Contracts List | Iris - Cardano Blockchain Explorer`;
   }, []);
 
   const columns: Column<Contracts>[] = [
@@ -116,7 +116,7 @@ const Transactions: React.FC = () => {
             total: fetchData.total,
             onChange: (page, size) => {
               history.replace({ search: stringify({ page, size }) });
-              mainRef.current?.scrollTo(0, 0);
+              mainRef.current?.scrollTo({ top: 0, behavior: "smooth" });
             }
           }}
         />