+++ src/pages/DelegatorLifecycle/index.tsx
@@ -33,6 +33,7 @@ import {
   Label,
   ReportButtonContainer
 } from "./styles";
+import { IReportLimit } from "../SPOLifecycle";
 
 interface Params {
   stakeId: string;
@@ -84,6 +85,7 @@ const DelegatorLifecycle = () => {
   const { data: listTabs, loading: loadingListTabs } = useFetch<ListStakeKeyResponse>(
     API.STAKE_LIFECYCLE.TABS(stakeId)
   );
+  const { data: dataReportLimit } = useFetch<IReportLimit>(API.REPORT.REPORT_LIMIT);
 
   useEffect(() => {
     if (listTabs && listTabs[tabsValid[tab]]) {
@@ -97,7 +99,7 @@ const DelegatorLifecycle = () => {
   }, [listTabs]);
 
   useEffect(() => {
-    document.title = `Staking Delegation Lifecycle ${stakeId} | Cardano Explorer`;
+    document.title = `Staking Delegation Lifecycle ${stakeId} | Iris - Cardano Blockchain Explorer`;
   }, [stakeId]);
 
   const changeMode = (mode: ViewMode) => {
@@ -133,9 +135,21 @@ const DelegatorLifecycle = () => {
                 </ButtonSwitch>
               </SwitchGroup>
             </BoxSwitchContainer>
-            <CustomTooltip title={!isLoggedIn ? "Please sign in to use this feature" : ""}>
+            <CustomTooltip
+              title={
+                !isLoggedIn
+                  ? "Please sign in to use this feature"
+                  : `Please note that you can only create ${
+                      dataReportLimit?.limitPer24hours || 0
+                    } reports within 24 hours`
+              }
+            >
               <ReportButtonContainer>
-                <ButtonReport disabled={!isLoggedIn} onClick={() => setOpen(true)} sidebar={+sidebar}>
+                <ButtonReport
+                  disabled={!isLoggedIn || dataReportLimit?.isLimitReached}
+                  onClick={() => setOpen(true)}
+                  sidebar={+sidebar}
+                >
                   Compose report
                 </ButtonReport>
               </ReportButtonContainer>