+++ src/pages/RegistrationPools/index.tsx
@@ -45,7 +45,7 @@ const RegistrationPools = () => {
 
   useEffect(() => {
     const title = poolType === POOL_TYPE.REGISTRATION ? "Registration" : "Deregistration";
-    document.title = `${title} Pools | Cardano Explorer`;
+    document.title = `${title} Pools | Iris - Cardano Blockchain Explorer`;
   }, [poolType]);
 
   const columns: Column<Registration>[] = [
@@ -151,7 +151,7 @@ const RegistrationPools = () => {
             ...pageInfo,
             onChange: (page, size) => {
               history.replace({ search: stringify({ page, size }) });
-              mainRef.current?.scrollTo(0, 0);
+              mainRef.current?.scrollTo({ top: 0, behavior: "smooth" });
             },
             total: fetchData.total
           }}